+++ ecs/src/Entity.cpp
@@ -8,7 +8,7 @@
 #include "Entity.hpp"
 
 namespace ecs {
-    size_t Entity::getId() const
+    uint32_t Entity::getId() const
     {
         return _id;
     }
